refactoring: Update GET near vehicles
