+++ localization/main.go
@@ -2,14 +2,13 @@ package main
 
 import (
 	"fmt"
-	grpc "localization/localization_grpc"
-	rest "localization/localization_rest"
+	"localization/server"
 )
 
 func main() {
 	fmt.Println("gRPC guild localization µ-service has been launched!")
 	// Start GIN server in a Goroutine
-	go rest.StartGINServer()
+	go server.StartGINServer()
 	// Start gRPC server in the main Goroutine
-	grpc.StartGRPCServer()
+	server.StartGRPCServer()
 }
